accepted changes from @stephenpasqualottounipd

+++ lib/image_helper/include/image_helper.h
@@ -2,7 +2,8 @@
 #ifndef image_helper_h
 #define image_helper_h
 
-/*
+
+  /*
  * This is the struct of image:
  *
  * matrice: [double[]] is the matrix of the image
@@ -10,18 +11,18 @@
  * intensity: [long int] is the intensity of the image
  * next: [struct immagine] is the pointer to the next image
  */
-struct immagine
-{
-	double matrice[784];
+struct immagine {
+	double matrice [784];
 	int label;
 	long int intensity;
 	struct immagine *next;
 };
 
+
 /*
  * This function return the value of the pixel at the position (x,y) of the image.
  */
-double pixel_at(int x, int y, const struct immagine *image);
+int pixel_at(int x, int y, const struct immagine *image);
 
 /*
  * This function set the value of the label with the value x.
@@ -36,26 +37,27 @@ int get_label(struct immagine *image);
 /*
  * This function return the value of the intensity of the image.
  */
-long int get_intensity(const struct immagine *image);
+long int get_intensity (const struct immagine *image);
 
 /*
  * This function print the image.
  */
-void print(const struct immagine *image);
+void print (const struct immagine *image);
 
 /*
  * This function return the distance between two images.
  */
-long int compute_distance(const struct immagine *image1, const struct immagine *image2);
+long int compute_distance (const struct immagine *image1, const struct immagine *image2);
 
 /*
  * This function return the image with the minimum distance from the image1.
  */
-char *compare_image(const struct immagine *image1, const struct immagine *image2);
+char *compare_image (const struct immagine *image1, const struct immagine *image2);
 
 /*
  * This function return 1 if the intensity of the image 1 is bigger than intensity of image2.
  */
-int compate_intensity(const struct immagine *image1, const struct immagine *image2);
+int compate_intensity (const struct immagine *image1, const struct immagine *image2);
+
 
 #endif
